talkclaw-agent/requirements.txt
livekit-agents[openai,deepgram,elevenlabs,silero]~=1.4
httpx~=0.27.0
python-dotenv~=1.0.0
